refactor: dark-mode to color-scheme helper
